requirements.txt
# Core framework
streamlit>=1.28.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# AI Integration
google-generativeai>=0.3.0

# Environment management
python-dotenv>=1.0.0

# File handling
openpyxl>=3.1.0

# Utilities
requests>=2.28.0
python-dateutil>=2.8.0

# Authentication and security
bcrypt>=4.0.0

# Configuration
pyyaml>=6.0

# Optional: For enhanced UI components
plotly>=5.15.0

reportlab

